fix: se refactorizo codigo

+++ Models/Dtos/Account-Profile-Dtos/CreateAccountProfileDto.cs
@@ -14,9 +14,9 @@ namespace AppVidaSana.Models.Dtos.Account_Profile_Dtos
 
         [JsonRequired] public string sex { get; set; } = null!;
 
-        [JsonRequired] public int stature { get; set; }
+        [JsonRequired] public float stature { get; set; }
 
-        [JsonRequired] public int weight { get; set; }
+        [JsonRequired] public float weight { get; set; }
 
         [JsonRequired] public string protocolToFollow { get; set; } = null!;
