+++ Models/Dtos/Account-Profile-Dtos/ReturnProfileDto.cs
@@ -10,9 +10,9 @@ namespace AppVidaSana.Models.Dtos.Cuenta_Perfil_Dtos
 
         [JsonRequired]  public string sex { get; set; } = null!;
 
-        [JsonRequired]  public int stature { get; set; }
+        [JsonRequired]  public float stature { get; set; }
 
-        [JsonRequired]  public int weight { get; set; }
+        [JsonRequired]  public float weight { get; set; }
 
         [JsonRequired] public string protocolToFollow { get; set; } = null!;
     }
